Run docker ps before SAM local run and move it off of the EDT (#1326)
* Run docker ps before SAM local run. It was only getting run before remote run but it should actually be run before both
* This has the bonus effect of moving docker ps off of the EDT which was causing IDE exceptions sometimes
