+++ .changes/next-release/bugfix-5f698882-d95f-47ca-b71e-6283b26054e1.json
@@ -1,0 +1,4 @@
+{
+  "type" : "bugfix",
+  "description" : "Fix an IDE error showing up during \"SAM local debug\" caused by running \"docker ps\" on the wrong thread"
+}
