+++ .changes/next-release/bugfix-d342bc1e-f088-4447-a0c4-7c34c04a00dd.json
@@ -1,0 +1,4 @@
+{
+  "type" : "bugfix",
+  "description" : "Fix being able to run \"SAM local run\" configurations without Docker running"
+}
