+++ resources/resources/software/aws/toolkits/resources/localized_messages.properties
@@ -306,7 +306,7 @@ lambda.debug.attach.error=Debugger attach error
 lambda.debug.attach.fail=Unable to attach to debugger process.
 lambda.debug.process.start.timeout=Timeout waiting for Debugger Worker process to start
 lambda.debug.waiting=Waiting for debugger...
-lambda.debug.docker.not_connected=Debugging Lambda locally requires Docker. Please make sure it is installed and running
+lambda.debug.docker.not_connected=Running/Debugging Lambdas locally requires Docker. Please make sure it is installed and running
 resource_selector.create=Create
 aws_connection.region.label=Region:
 aws_connection.credentials.label=Credentials:
